Create ProjectFundingFilesAttachmentDialog
