+++ moped-editor/src/views/projects/projectView/DeleteConfirmationModal.js
@@ -15,6 +15,7 @@ const DeleteConfirmationModal = ({
   setIsDeleteConfirmationOpen,
   children,
   confirmationText,
+  additionalConfirmationText,
   actionButtonText = "Delete",
   actionButtonIcon,
 }) => {
@@ -22,7 +23,7 @@ const DeleteConfirmationModal = ({
     setIsDeleteConfirmationOpen(false);
   };
 
-  const defaultConfirmationText = `Are you sure you want to ${actionButtonText.toLowerCase()} this ${type}?`;
+  const defaultConfirmationText = `Are you sure you want to ${actionButtonText.toLowerCase()} this ${type}?${additionalConfirmationText ? ` ${additionalConfirmationText}` : ""}`;
   const ActionIcon = actionButtonIcon || <DeleteIcon />;
 
   return (
